+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+* **0.1.3**
+  * Implemented a Slider for the song Position. Can be used to search in the song
+  * Fixed a bug with the "jumpTo" function
+
 * **0.1.2**
   * removed autopublish package